refactor: :label: move m3u8 types
